Make replication task convertor injectable by fx
