+++ service/history/replication/raw_task_converter.go
@@ -42,36 +42,80 @@ import (
 	"go.temporal.io/server/common/namespace"
 	"go.temporal.io/server/common/persistence"
 	"go.temporal.io/server/common/persistence/versionhistory"
+	"go.temporal.io/server/service/history/shard"
 	"go.temporal.io/server/service/history/tasks"
 	"go.temporal.io/server/service/history/workflow"
 	wcache "go.temporal.io/server/service/history/workflow/cache"
 )
 
-func convertTask(
-	ctx context.Context,
-	task tasks.Task,
-	shardID int32,
-	workflowCache wcache.Cache,
-	executionManager persistence.ExecutionManager,
-	logger log.Logger,
-) (*replicationspb.ReplicationTask, error) {
-	switch task := task.(type) {
-	case *tasks.SyncActivityTask:
-		return convertActivityStateReplicationTask(ctx, task, workflowCache)
-	case *tasks.SyncWorkflowStateTask:
-		return convertWorkflowStateReplicationTask(ctx, task, workflowCache)
-	case *tasks.HistoryReplicationTask:
-		return convertHistoryReplicationTask(
-			ctx,
-			task,
-			shardID,
-			workflowCache,
-			executionManager,
-			logger,
-		)
-	default:
-		return nil, errUnknownReplicationTask
+type (
+	SourceTaskConvertorImpl struct {
+		historyEngine           shard.Engine
+		namespaceCache          namespace.Registry
+		clientClusterShardCount int32
+		clientClusterName       string
+		clientShardKey          ClusterShardKey
 	}
+	SourceTaskConvertor interface {
+		Convert(task tasks.Task) (*replicationspb.ReplicationTask, error)
+	}
+	SourceTaskConvertorFactory func(
+		historyEngine shard.Engine,
+		shardContext shard.Context,
+		clientClusterShardCount int32,
+		clientClusterName string,
+		clientShardKey ClusterShardKey,
+	) SourceTaskConvertor
+)
+
+func NewSourceTaskConvertor(
+	historyEngine shard.Engine,
+	namespaceCache namespace.Registry,
+	clientClusterShardCount int32,
+	clientClusterName string,
+	clientShardKey ClusterShardKey,
+) *SourceTaskConvertorImpl {
+	return &SourceTaskConvertorImpl{
+		historyEngine:           historyEngine,
+		namespaceCache:          namespaceCache,
+		clientClusterShardCount: clientClusterShardCount,
+		clientClusterName:       clientClusterName,
+		clientShardKey:          clientShardKey,
+	}
+}
+
+func (c *SourceTaskConvertorImpl) Convert(
+	task tasks.Task,
+) (*replicationspb.ReplicationTask, error) {
+	if namespaceEntry, err := c.namespaceCache.GetNamespaceByID(
+		namespace.ID(task.GetNamespaceID()),
+	); err == nil {
+		shouldProcessTask := false
+	FilterLoop:
+		for _, targetCluster := range namespaceEntry.ClusterNames() {
+			if c.clientClusterName == targetCluster {
+				shouldProcessTask = true
+				break FilterLoop
+			}
+		}
+		if !shouldProcessTask {
+			return nil, nil
+		}
+	}
+	// if there is error, then blindly send the task, better safe than sorry
+
+	clientShardID := common.WorkflowIDToHistoryShard(task.GetNamespaceID(), task.GetWorkflowID(), c.clientClusterShardCount)
+	if clientShardID != c.clientShardKey.ShardID {
+		return nil, nil
+	}
+
+	ctx, cancel := context.WithTimeout(context.Background(), replicationTimeout)
+	defer cancel()
+	replicationTask, err := c.historyEngine.ConvertReplicationTask(ctx, task)
+	if err != nil {
+		return nil, err
+	}
+	return replicationTask, nil
 }
 
 func convertActivityStateReplicationTask(
